-fix, sobre las responses de las apis y se agregan mejor  manejo de los mappers para spring boot

+++ src/main/java/dev/andreszapata/bankfuse/domain/model/Product.java
@@ -13,7 +13,7 @@ public class Product {
 
     private Long idProduct;
     private TipoCuenta tipoCuenta;
-    private int numeroCuenta;
+    private Long numeroCuenta;
     private EstadoCuenta estadoCuenta;
     private Double saldo;
     private Boolean excentaGmf;
